+++ inst/doc/a30-chunk-engine.R
@@ -1,0 +1,31 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ---- message=FALSE, echo=FALSE-----------------------------------------------
+library(caracas)
+
+## ---- include = FALSE---------------------------------------------------------
+inline_code <- function(x) {
+  x
+}
+
+if (!has_sympy()) {
+  # SymPy not available, so the chunks shall not be evaluated
+  knitr::opts_chunk$set(eval = FALSE)
+  
+  inline_code <- function(x) {
+    deparse(substitute(x))
+  }
+}
+
+## -----------------------------------------------------------------------------
+tex <- function(x) {
+  caracas::tex(x, zero_as_dot = TRUE)
+}
+
+## ---- include = FALSE---------------------------------------------------------
+rm(tex)
+